General upkeep
Added attribution in the comments. Comments are very basic but most
of the code is fairly self explanatory. The difficult part of the
assignment was designing the grammar not implementing it. In addition
some minor changes were made to part 4. Currently part 4 does not generate
the appropriate machine code as it does not have any sort of while function
just a series of code chunks.